apps/core/auth-service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
orjson==3.11.5
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.21
pyyaml==6.0.3
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
sentry-sdk==2.48.0
shellingham==1.5.4
starlette==0.50.0
tomli==2.3.0
typer==0.21.0
typing-extensions==4.15.0
typing-inspection==0.4.2
ujson==5.11.0
urllib3==2.6.2
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
